bern/requirements.txt
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.8.1
beautifulsoup4==4.11.2
certifi==2022.5.18.1
charset-normalizer==2.0.12
colorama==0.4.6
distro==1.9.0
Django==5.2
django-braces==1.15.0
django-media-pil==1.0.0
django-parler==2.3
django-paypal==2.0
et-xmlfile==1.1.0
exceptiongroup==1.3.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.3
jiter==0.12.0
lxml==4.9.2
numpy==1.25.1
openai==2.16.0
openpyxl==3.1.2
openpyxl-image-loader==1.0.5
pandas==2.0.3
pgvector==0.4.2
Pillow==9.1.1
psycopg2 @ file:///C:/ci_310/psycopg2_1642085998310/work
psycopg2-binary==2.9.11
pycrop==1.0.1
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.8.2
pytz @ file:///C:/Windows/TEMP/abs_90eacd4e-8eff-491e-b26e-f707eba2cbe1ujvbhqz1/croots/recipe/pytz_1654762631027/work
requests==2.28.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.4
sqlparse @ file:///tmp/build/80754af9/sqlparse_1602184451250/work
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2023.3
urllib3==1.26.9
wincertstore==0.2
xlrd==2.0.1
